requirements.txt
PyGithub>=1.59.0
pandas>=2.0.0
pytest>=7.0.0
vcrpy>=6.0.0